+++ _inc/lib/core-api/class.jetpack-core-api-module-endpoints.php
@@ -572,6 +572,7 @@ class Jetpack_Core_API_Data extends Jetpack_Core_API_XMLRPC_Consumer_Endpoint {
 				case 'google':
 				case 'bing':
 				case 'pinterest':
+				case 'yandex':
 					$grouped_options          = $grouped_options_current = (array) get_option( 'verification_services_codes' );
 					$grouped_options[$option] = $value;
 
@@ -580,6 +581,10 @@ class Jetpack_Core_API_Data extends Jetpack_Core_API_XMLRPC_Consumer_Endpoint {
 					break;
 
 				case 'sharing_services':
+					if ( ! class_exists( 'Sharing_Service' ) && ! @include( JETPACK__PLUGIN_DIR . 'modules/sharedaddy/sharing-service.php' ) ) {
+						break;
+					}
+
 					$sharer = new Sharing_Service();
 
 					// If option value was the same, consider it done.
@@ -589,14 +594,22 @@ class Jetpack_Core_API_Data extends Jetpack_Core_API_XMLRPC_Consumer_Endpoint {
 				case 'button_style':
 				case 'sharing_label':
 				case 'show':
-					$sharer                   = new Sharing_Service();
-					$grouped_options          = $sharer->get_global_options();
-					$grouped_options[$option] = $value;
-					$updated                  = $sharer->set_global_options( $grouped_options );
+					if ( ! class_exists( 'Sharing_Service' ) && ! @include( JETPACK__PLUGIN_DIR . 'modules/sharedaddy/sharing-service.php' ) ) {
+						break;
+					}
+
+					$sharer = new Sharing_Service();
+					$grouped_options = $sharer->get_global_options();
+					$grouped_options[ $option ] = $value;
+					$updated = $sharer->set_global_options( $grouped_options );
 					break;
 
 				case 'custom':
-					$sharer  = new Sharing_Service();
+					if ( ! class_exists( 'Sharing_Service' ) && ! @include( JETPACK__PLUGIN_DIR . 'modules/sharedaddy/sharing-service.php' ) ) {
+						break;
+					}
+
+					$sharer = new Sharing_Service();
 					$updated = $sharer->new_service( stripslashes( $value['sharing_name'] ), stripslashes( $value['sharing_url'] ), stripslashes( $value['sharing_icon'] ) );
 
 					// Return new custom service
@@ -604,7 +617,11 @@ class Jetpack_Core_API_Data extends Jetpack_Core_API_XMLRPC_Consumer_Endpoint {
 					break;
 
 				case 'sharing_delete_service':
-					$sharer  = new Sharing_Service();
+					if ( ! class_exists( 'Sharing_Service' ) && ! @include( JETPACK__PLUGIN_DIR . 'modules/sharedaddy/sharing-service.php' ) ) {
+						break;
+					}
+
+					$sharer = new Sharing_Service();
 					$updated = $sharer->delete_service( $value );
 					break;
 
@@ -1005,6 +1022,9 @@ class Jetpack_Core_API_Module_Data_Endpoint {
 						break;
 					case 'pinterest':
 						$services[] = 'Pinterest';
+						break;
+					case 'yandex':
+						$services[] = 'Yandex';
 						break;
 				}
 			}
